CCCP-227, refactor: squash bifrostutils.py & utils.py
